% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{embryo_celltype_colours}
\alias{embryo_celltype_colours}
\title{Celltype colours from Pijuan-Sala et al.}
\format{A vector of hexcodes named according to the appropriate
celltype; celltypes match those in the metadata.}
\usage{
embryo_celltype_colours
}
\description{
A vector containing the colour hexcodes that were used
in Pijuan-Sala et al., for quick plotting of the data.
}
\references{
Pijuan-Sala B, Griffiths JA, Guibentif C et al. (2019).
A single-cell molecular map of mouse gastrulation and early organogenesis.
\emph{Nature} 566, 7745:490-495.
}
\keyword{datasets}
