\name{MouseGastrulationData-package}
\alias{MouseGastrulationData-package}
\alias{MouseGastrulationData}

\docType{package}
\title{
\packageTitle{MouseGastrulationData}
}
\description{
\packageDescription{MouseGastrulationData}
}
\details{

This package contains the processed 10X Genomics data from Pijuan-Sala et al. (2019), Pijuan-Sala et al. (2020), Guibentif et al. (2020) and Lohoff et al. (2020.).

The data were processed as described in the methods that accompany the paper.

}
\author{
\packageAuthor{MouseGastrulationData}

Maintainer: \packageMaintainer{MouseGastrulationData}
}

\references{

Blanca Pijuan-Sala*, Jonathan A. Griffiths*, Carolina Guibentif*, Tom W. Hiscock, Wajid Jawaid, Fernando J. Calero-Nieto, Carla Mulas, Ximena Ibarra-Soria, Richard C.V. Tyser, Debbie Lee Lian Ho, Wolf Reik, Shankar Srinivas, Benjamin D. Simons, Jennifer Nichols, John C. Marioni, Berthold Göttgens. A single-cell molecular map of mouse gastrulation and early organogenesis. Nature, 566, pp490-495 (2019).

Blanca Pijuan-Sala, Nicola K. Wilson, Jun Xia, Xiaomeng Hou, Rebecca L. Hannah, Sarah Kinston, Fernando J. Calero-Nieto, Olivier Poirion, Sebastian Preissl, Feng Liu, Berthold Göttgens. Single-cell chromatin accessibility maps reveal regulatory programs driving early mouse organogenesis. Nat Cell Biol, 22, 487–497 (2020).

Carolina Guibentif*, Jonathan A. Griffiths*, Ivan Imaz-Rosshandler, Shila Ghazanfar, Jennifer Nichols, Valerie Wilson, Berthold Göttgens, John C. Marioni. Diverse Routes toward Early Somites in the Mouse Embryo. Dev Cell. 2020 Dec 1:S1534-5807(20)30889-3.

T. Lohoff*, S. Ghazanfar*, A. Missarova, N. Koulena, N. Pierson, J.A. Griffiths, E.S. Bardot, C.-H.L. Eng, R.C.V. Tyser, R. Argelaguet, C. Guibentif, S. Srinivas, J. Briscoe, B.D. Simons, A.-K. Hadjantonakis, B. Göttgens, W. Reik, J. Nichols, L. Cai, J.C. Marioni. Highly multiplexed spatially resolved gene expression profiling of mouse organogenesis. bioRxiv 2020.11.20.391896.
}
