\name{MouseGastrulationData-package}
\alias{MouseGastrulationData-package}
\alias{MouseGastrulationData}

\docType{package}
\title{
\packageTitle{MouseGastrulationData}
}
\description{
\packageDescription{MouseGastrulationData}
}
\details{

This package contains the processed 10X Genomics data from Pijuan-Sala et al. (2019).

The data were processed as described in the methods that accompany the paper.

}
\author{
\packageAuthor{MouseGastrulationData}

Maintainer: \packageMaintainer{MouseGastrulationData}
}

\references{

Blanca Pijuan-Sala*, Jonathan A. Griffiths*, Carolina Guibentif*, Tom W. Hiscock, Wajid Jawaid, Fernando J. Calero-Nieto, Carla Mulas, Ximena Ibarra-Soria, Richard C.V. Tyser, Debbie Lee Lian Ho, Wolf Reik, Shankar Srinivas, Benjamin D. Simons, Jennifer Nichols, John C. Marioni, Berthold Göttgens. A single-cell molecular map of mouse gastrulation and early organogenesis. Nature, 566, pp490-495 (2019).
}
